common baseRun function, avoid copy-paste
